requirements.txt
numpy>=1.24
scipy>=1.10
pandas>=2.0
matplotlib>=3.8
seaborn>=0.13
click>=8.1
tqdm>=4.66
pytest>=7.4
scikit-learn>=1.5
joblib>=1.3
pyyaml>=6.0
h5py>=3.10

